export storage interfaces
