+++ packages/chamber-core/index.js
@@ -1,5 +1,9 @@
 const {
-  BaseWallet
+  BaseWallet,
+  StorageInterface,
+  BigStorageInterface,
+  MockStorage,
+  MockBigStorage
 } = require('./lib/wallet')
 const Block = require('./lib/block')
 const Merkle = require('./lib/smt')
@@ -12,6 +16,10 @@ const BufferUtils = require('./lib/BufferUtils')
 
 module.exports = {
   BaseWallet,
+  StorageInterface,
+  BigStorageInterface,
+  MockStorage,
+  MockBigStorage,
   Block,
   BufferUtils,
   Merkle,
