+++ packages/chamber-core/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@cryptoeconomicslab/chamber-core",
-  "version": "1.0.2",
+  "version": "1.0.3",
   "description": "",
   "main": "index.js",
   "scripts": {
