+++ packages/chamber-core/src/wallet/index.js
@@ -1,0 +1,17 @@
+const {
+  StorageInterface,
+  BigStorageInterface,
+  MockStorage,
+  MockBigStorage
+} = require('./storage')
+const {
+  BaseWallet
+} = require('./wallet')
+
+module.exports = {
+  BaseWallet,
+  StorageInterface,
+  BigStorageInterface,
+  MockStorage,
+  MockBigStorage
+}
